1.0.0
建立用到的主要类

+++ src/main/matter/Point.java
@@ -1,29 +1,29 @@
 package main.matter;
 
 public class Point {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
 	Point(){
 		this.x = 0;
 		this.y = 0;
 	}
 	
-	Point(int x,int y){
+	Point(double x,double y){
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
